LayoutTests/java/lc3/JavaObject/JavaObjectToByte-003-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject to byte

FAIL dt.PUB_DOUBLE_REPRESENTATION = 128;dt.setByte( dt ); dt.PUB_BYTE should throw an exception. Was -128.
FAIL dt.PUB_DOUBLE_REPRESENTATION = 128;dt.setByte( dt ); dt.getByte() should throw an exception. Was -128.
PASS typeof dt.getByte() is 'number'
PASS successfullyParsed is true

TEST COMPLETE
 
